+++ test/cases/children-and-slots/src/components/paragraph.js
